+++ pojo-lens/src/main/java/laughing/man/commits/report/SavedReportKind.java
@@ -1,0 +1,9 @@
+package laughing.man.commits.report;
+
+/**
+ * Identifies the query language used in a {@link SavedReport}.
+ */
+public enum SavedReportKind {
+    SQL_LIKE,
+    NATURAL
+}